gitserver: Convert MergeBase client call to server-side
This PR is a PoC for what it could look like to convert the client-side implementation of common git functionality into well-scoped server-side offered APIs, starting with a not very widely used one, MergeBase.

I did not implement an HTTP equivalent because only next-minor clients will use it, which means that gRPC will be on by default by then.
The MergeBase endpoint was so rarely changed that I don't expect that to cost any velocity in backports.